feat: unified tool registry with tier classification and arg validation
Introduces tool-registry.ts as the single source of truth for all ORION
management tools. All 19 tools registered with tier (read/write/destructive),
parallelSafe flag, and availableIn (task/chat/both) classifications.

management-tools.ts re-exports from registry for full backwards compatibility.

Adds validateToolArgs() — checks required fields and types against inputSchema
before any tool execution. Wired into openai-runner, ollama-runner, and the
claude.ts chat path; returns a structured error to the LLM on validation
failure instead of passing bad args to execFile.
